refactor and pull out http
